# blockchain: Use skip list for ancestor traversal.