There are several places when processing blocks and headers that require the ability to traverse the chain to ancestors of a given block such as tallying votes and versions to determine threshold states, finding forks, and calculating PoW difficulty, ticket price, and sequence locks.

Currently most traversal of this type is linear and therefore does not scale well as the chain height grows.  In addition, the ability to quickly locate ancestors, potentially deep in history, will be needed when decoupling the connection code from the download logic as well as several planned future optimizations.

Consequently, this significantly optimizes ancestor traversal by introducing a deterministic skip list and adds tests to ensure proper functionality.

The following is a before and after comparison of ancestor traversal for a large number of nodes:

```
benchmark           old ns/op     new ns/op   delta
------------------------------------------------------
BenchmarkAncestor   67901581      158         -100.00%

benchmark           old allocs   new allocs   delta
------------------------------------------------------
BenchmarkAncestor   0            0            +0.00%

benchmark           old bytes    new bytes    delta
------------------------------------------------------
BenchmarkAncestor   0            0            +0.00%
```

The following is a scatter plot of the number of steps to traverse back to a height of 1 for all heights up to 2^14 along with a logarithmic regression to help visualize the behavior.

![steps_vs_height](https://user-images.githubusercontent.com/2115102/69477340-aad66480-0daa-11ea-812d-7563be10bd03.png)